refactoring xqa
